init Sensors in Factory
